feat: Add Streamlit UI components for agent service chat interface
- Implemented main UI structure in `main_page.py` to wire agent client, sidebar, message stream, and feedback widget.
- Created reusable UI components in `components` directory, including message rendering and sidebar controls.
- Introduced `draw_message.py` for rendering chat messages with support for streaming and tool calls.
- Added `side_panel.py` for chat controls and settings configuration.
- Developed utility functions in `utils/message.py` for handling messages and feedback.
- Updated constants in `utils/constants.py` for static UI configuration.
- Enhanced message handling to support multimodal inputs and feedback recording.
- Updated version to 0.9.1 in `pyproject.toml` and adjusted dependencies in `uv.lock`.
- Added tests for new features and ensured compatibility with existing functionality.
