+++ .env.example
@@ -113,4 +113,10 @@ MODEL_CONFIGS_PATH=
 # Amazon Bedrock Knowledge Base ID
 AWS_KB_ID=
 
+# Detect/resume interrupted runs (human-in-the-loop). Default true; set false to skip the per-request
+# checkpointer read when no agent uses interrupt().
 CHECK_INTERRUPTS=
+
+# Max non-text attachments (image/file/audio/video) per message. Empty = no toolkit limit
+# (the model provider still enforces its own per-model caps).
+MULTIMODAL_MAX_ATTACHMENTS=
