+++ langgraph_agent_toolkit/ui/utils/constants.py
@@ -1,0 +1,7 @@
+"""Static UI configuration: titles, icons, and accepted upload types."""
+
+APP_TITLE = "Agent Service Toolkit"
+APP_ICON = "🧰"
+
+# File types accepted for multimodal chat uploads, mapped to LangChain content blocks.
+MULTIMODAL_FILE_TYPES = ["png", "jpg", "jpeg", "gif", "webp", "pdf", "wav", "mp3", "m4a", "mp4", "webm"]
